+++ ext/polars/src/lib.rs
@@ -552,6 +552,7 @@ fn init(ruby: &Ruby) -> RbResult<()> {
     // maybe add to different class
     let class = module.define_module("Plr")?;
     class.define_singleton_method("dtype_cols", function!(functions::lazy::dtype_cols2, 1))?;
+    class.define_singleton_method("index_cols", function!(functions::lazy::index_cols, 1))?;
     class.define_singleton_method("col", function!(functions::lazy::col, 1))?;
     class.define_singleton_method("len", function!(functions::lazy::len, 0))?;
     class.define_singleton_method("first", function!(functions::lazy::first, 0))?;